Fix errors from previous commit
-

- remove idea folder

- refactored code
- spearated out models from the main code
- changed the template
- added some more duplication avoiding logic
- added persistence
