+++ examples/core/apidoc.html.json
@@ -1,0 +1,1 @@
+{"ApiSpecs":[{"HttpVerb":"GET","Path":"/","Calls":[{"Id":0,"CurrentPath":"/","MethodType":"GET","PostForm":null,"RequestHeader":{"Content-Type":" application/json\r"},"CommonRequestHeaders":null,"ResponseHeader":{},"RequestUrlParams":{},"RequestBody":"","ResponseBody":"Hi there, I love !","ResponseCode":200}]},{"HttpVerb":"POST","Path":"/say_it","Calls":[{"Id":1,"CurrentPath":"/say_it","MethodType":"POST","PostForm":null,"RequestHeader":{"Content-Type":" application/json\r"},"CommonRequestHeaders":null,"ResponseHeader":{"Test":"tesasasdasd"},"RequestUrlParams":{},"RequestBody":"","ResponseBody":"","ResponseCode":200}]}]}
