+++ examples/core/server.go
@@ -26,7 +26,7 @@ func postHandler(w http.ResponseWriter, r *http.Request) {
 }
 
 func main() {
-	yaag.Init(&yaag.Config{On: true, DocTitle: "Core", DocPath: "apidoc.html"})
+	yaag.Init(&yaag.Config{On: true, DocTitle: "Core", DocPath: "apidoc.html", BaseUrls: map[string]string{"Production": "http://testing.com"}})
 	http.HandleFunc("/", middleware.HandleFunc(handler))
 	http.HandleFunc("/say_it", middleware.HandleFunc(postHandler))
 	http.ListenAndServe(":8080", nil)
