+++ examples/martini/server.go
@@ -7,7 +7,7 @@ import (
 )
 
 func main() {
-	yaag.Init(&yaag.Config{On: true, DocTitle: "Gorilla Mux", DocPath: "apidoc.html"})
+	yaag.Init(&yaag.Config{On: true, DocTitle: "Gorilla Mux", DocPath: "apidoc.html", BaseUrls: map[string]string{"Production": "", "Staging": ""}})
 	m := martini.Classic()
 	m.Use(martiniyaag.Document)
 	m.Get("/", func() string {
