+++ yaag/config.go
@@ -1,0 +1,10 @@
+package yaag
+
+type Config struct {
+	On bool
+
+	BaseUrls map[string]string
+
+	DocTitle string
+	DocPath  string
+}
