+++ yaag/constants.go
@@ -1,0 +1,13 @@
+package yaag
+
+var CommonHeaders = []string{
+	"Accept",
+	"Accept-Encoding",
+	"Accept-Language",
+	"Cache-Control",
+	"Content-Length",
+	"Content-Type",
+	"Origin",
+	"User-Agent",
+	"X-Forwarded-For",
+}
