+++ yaag/models/apicall.go
@@ -1,0 +1,19 @@
+package models
+
+type ApiCall struct {
+	Id int
+
+	CurrentPath string
+	MethodType  string
+
+	PostForm map[string]string
+
+	RequestHeader        map[string]string
+	CommonRequestHeaders map[string]string
+	ResponseHeader       map[string]string
+	RequestUrlParams     map[string]string
+
+	RequestBody  string
+	ResponseBody string
+	ResponseCode int
+}
