+++ yaag/models/apispec.go
@@ -1,0 +1,7 @@
+package models
+
+type ApiSpec struct {
+	HttpVerb string
+	Path     string
+	Calls    []ApiCall
+}
