+++ yaag/models/spec.go
@@ -1,0 +1,5 @@
+package models
+
+type Spec struct {
+	ApiSpecs []ApiSpec
+}
